Added snmp configurator
