+++ cloudshell/networking/juniper/cli/juniper_command_modes.py
@@ -4,7 +4,7 @@
 from collections import OrderedDict
 
 from cloudshell.cli.service.command_mode import CommandMode
-from cloudshell.shell.core.api_utils import decrypt_password
+from cloudshell.shell.standards.core import decrypt_password
 
 
 # class CliCommandMode(CommandMode):
